third_party/WebKit/LayoutTests/http/tests/inspector/network/network-disable-cache-cors.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script type="text/javascript">
function sendXHR(callback)
{
    var xhr = new XMLHttpRequest();
    xhr.onreadystatechange = function()
    {
        if (xhr.readyState === XMLHttpRequest.DONE)
            callback();
    };
    xhr.open("POST", "http://localhost:8000/inspector/network/resources/cors.cgi");
    xhr.setRequestHeader("Content-Type", "application/xml");
    xhr.send("<xml></xml>");
}


function step1()
{
    sendXHR(step2); // Should issue OPTIONS and POST requests.
}

function step2()
{
    sendXHR(step3); // Should issue OPTIONS and POST requests.
}

function step3()
{
    console.log("Done.");
}

function test()
{
    NetworkTestRunner.makeFetch("http://localhost:8080/inspector/network/resources/cors-redirect.cgi", {headers: {"x-test": "redirect"}}, function () {});
    NetworkTestRunner.makeFetch("http://localhost:8080/inspector/network/resources/cors-redirect.cgi", {}, disableCache);

    async function disableCache()
    {
        await TestRunner.NetworkAgent.setCacheDisabled(true);
        NetworkTestRunner.recordNetwork();
        ConsoleTestRunner.addConsoleSniffer(step4);
        TestRunner.evaluateInPage("step1();");
    }

    async function step4()
    {
        await TestRunner.NetworkAgent.setCacheDisabled(false);
        var requests = NetworkTestRunner.networkRequests();
        for (var i = 0; i < requests.length; ++i) {
            var request = requests[i];
            var method = request.requestMethod;
            if (method === "OPTIONS" || method === "POST")
                TestRunner.addResult(method + ": " + request.url());
        }
        TestRunner.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests that preflight OPTIONS is always sent if 'Disable cache' is checked, and that network instrumentation does not produce errors for redirected preflights.</p>
</body>
</html>
